@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -17,20 +17,28 @@ let package = Package(
             targets: ["Toggles"])
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0")
+        .package(url: "https://github.com/apple/swift-docc-plugin", .upToNextMinor(from: "1.4.3"))
     ],
     targets: [
         .target(
             name: "Toggles",
             dependencies: [],
             path: "Sources",
-            resources: [.process("Resources")]
+            resources: [.process("Resources")],
+            swiftSettings: [
+                .enableExperimentalFeature("InternalImportsByDefault"),
+                .enableExperimentalFeature("ExistentialAny")
+            ]
         ),
         .testTarget(
             name: "TogglesTests",
             dependencies: ["Toggles"],
             path: "Tests",
-            resources: [.process("Resources")]
+            resources: [.process("Resources")],
+            swiftSettings: [
+                .enableExperimentalFeature("InternalImportsByDefault"),
+                .enableExperimentalFeature("ExistentialAny")
+            ]
         )
     ]
 )
