@@ -1,4 +1,4 @@
-// swift-tools-version: 5.9
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -14,7 +14,7 @@ let package = Package(
                     targets: ["ToggleGen"])
     ],
     dependencies: [
-        .package(url: "https://github.com/apple/swift-argument-parser.git", .upToNextMajor(from: "1.0.0")),
+        .package(url: "https://github.com/apple/swift-argument-parser.git", .upToNextMinor(from: "1.5.0")),
         .package(url: "https://github.com/stencilproject/Stencil.git", exact: "0.15.1"),
     ],
     targets: [
@@ -24,11 +24,21 @@ let package = Package(
                 .product(name: "ArgumentParser", package: "swift-argument-parser"),
                 .product(name: "Stencil", package: "Stencil"),
             ],
-            path: "Sources"),
+            path: "Sources",
+            swiftSettings: [
+                .enableExperimentalFeature("InternalImportsByDefault"),
+                .enableExperimentalFeature("ExistentialAny")
+            ]
+        ),
         .testTarget(
             name: "ToggleGenTests",
             dependencies: ["ToggleGen"],
             path: "Tests",
-            resources: [.process("Resources")])
+            resources: [.process("Resources")],
+            swiftSettings: [
+                .enableExperimentalFeature("InternalImportsByDefault"),
+                .enableExperimentalFeature("ExistentialAny")
+            ]
+        )
     ]
 )
