@@ -3,7 +3,7 @@
 import Foundation
 
 /// Enumeration of supported cipher algorithms used to encrypt and decrypt secure toggles.
-public enum Algorithm: String {
+public enum Algorithm: String, Sendable {
     /// The [ChaChaPoly cipher](https://developer.apple.com/documentation/cryptokit/chachapoly) from Apple's CryptoKit.
     /// See [Wikipedia reference](https://en.wikipedia.org/wiki/ChaCha20-Poly1305) for more details.
     case chaCha20Poly1305
@@ -13,7 +13,7 @@ public enum Algorithm: String {
 public typealias CipherKey = String
 
 /// Configuration describing what cipher algorigthm use for encrypting and decrypting secure toggles.
-public struct CipherConfiguration {
+public struct CipherConfiguration: Sendable {
     let algorithm: Algorithm
     let key: CipherKey
     let ignoreEmptyStrings: Bool
